outsourced validation source code to lib-folder
